backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
Pillow==9.0.0
django-bootstrap5==22.2
djangorestframework-simplejwt==4.7.2
pytest-pythonpath==0.7.3
pytest-django==4.4.0
django-filter==23.1
gunicorn==20.1.0
